test(core): fix claude-code-official tier assertions + split GitHubIndexer.test.ts (SMI-5205)
- Update 3 test files: claude-code-official topic now assigns 'official' tier
  (was 'verified') per Wave 1 GitHubIndexer change
- Split GitHubIndexer.test.ts (645 lines → 414) into GitHubIndexer.repository.test.ts
  (242 lines) to satisfy 500-line gate; IndexerRepository + integration tests moved
